feat(sdk-verification)!: Implementing DIF Presentation exchange protocol 2.0 for SDK to SDK OOB verification (#196)
Signed-off-by: Pete Vielhaber <peter.vielhaber@iohk.io>
Co-authored-by: Pete Vielhaber <peter.vielhaber@iohk.io>

BREAKING CHANGES:
1. JWTVerifiable Credential
Removed fromJWT which was used internally most of the times, has been replaced by fromJWS which just needs the JWS string.
2. Agent Credentials
This class is also for internal usee and its constructor has changed, now requiring an instance of mercury and did higher functions.

Most implementations will not be affected by this breaking change but could be in more advanced implementations.
3. Presentation Request
The type field in the constructor now needs to be called with enum AttachmentFormats instead of CredentialType, making a clear gap between didcomm message format types and credential types.